+++ saleor/graphql/product/mutations/products.py
@@ -25,8 +25,8 @@ from ....order.tasks import recalculate_orders_task
 from ....product import ProductMediaTypes, ProductTypeKind, models
 from ....product.error_codes import CollectionErrorCode, ProductErrorCode
 from ....product.search import (
-    update_product_search_document,
-    update_products_search_document,
+    update_product_search_vector,
+    update_products_search_vector,
 )
 from ....product.tasks import (
     update_product_discounted_price_task,
@@ -676,7 +676,7 @@ class ProductCreate(ModelMutation):
     @classmethod
     def post_save_action(cls, info, instance, _cleaned_input):
         product = models.Product.objects.prefetched_for_webhook().get(pk=instance.pk)
-        update_product_search_document(instance)
+        update_product_search_vector(instance)
         info.context.plugins.product_created(product)
 
     @classmethod
@@ -729,7 +729,7 @@ class ProductUpdate(ProductCreate):
     @classmethod
     def post_save_action(cls, info, instance, _cleaned_input):
         product = models.Product.objects.prefetched_for_webhook().get(pk=instance.pk)
-        update_product_search_document(instance)
+        update_product_search_vector(instance)
         info.context.plugins.product_updated(product)
 
 
@@ -1057,7 +1057,7 @@ class ProductVariantCreate(ModelMutation):
             AttributeAssignmentMixin.save(instance, attributes)
 
         generate_and_set_variant_name(instance, cleaned_input.get("sku"))
-        update_product_search_document(instance.product)
+        update_product_search_vector(instance.product)
         event_to_call = (
             info.context.plugins.product_variant_created
             if new_variant
@@ -1153,7 +1153,7 @@ class ProductVariantDelete(ModelDeleteMutation):
         # Update the "discounted_prices" of the parent product
         update_product_discounted_price_task.delay(instance.product_id)
         product = models.Product.objects.get(id=instance.product_id)
-        update_product_search_document(product)
+        update_product_search_vector(product)
         # if the product default variant has been removed set the new one
         if not product.default_variant:
             product.default_variant = product.variants.first()
@@ -1394,7 +1394,7 @@ class ProductTypeUpdate(ProductTypeCreate):
             or "variant_attributes" in cleaned_input
         ):
             products = models.Product.objects.filter(product_type=instance)
-            update_products_search_document(products)
+            update_products_search_vector(products)
 
 
 class ProductTypeDelete(ModelDeleteMutation):